pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "python-frontmatter>=1.1.0",
]
